Move protocol-related data types to protocol/ dir
