+++ packages/liveblocks-core/src/LiveList.ts
@@ -5,10 +5,10 @@ import { comparePosition, makePosition } from "./lib/position";
 import { LiveRegister } from "./LiveRegister";
 import type { LiveNode, Lson } from "./types/Lson";
 import type { ParentToChildNodeMap } from "./types/NodeMap";
-import type { CreateChildOp, CreateListOp, CreateOp, Op } from "./types/Op";
-import { OpCode } from "./types/Op";
-import type { IdTuple, SerializedList } from "./types/SerializedCrdt";
-import { CrdtType } from "./types/SerializedCrdt";
+import type { CreateChildOp, CreateListOp, CreateOp, Op } from "./protocol/Op";
+import { OpCode } from "./protocol/Op";
+import type { IdTuple, SerializedList } from "./protocol/SerializedCrdt";
+import { CrdtType } from "./protocol/SerializedCrdt";
 import type {
   LiveListUpdateDelta,
   LiveListUpdates,
