+++ packages/liveblocks-core/src/__tests__/LiveList.test.ts
@@ -2,9 +2,9 @@ import { RoomScope } from "../AuthToken";
 import { LiveList } from "../LiveList";
 import { LiveMap } from "../LiveMap";
 import { LiveObject } from "../LiveObject";
-import { OpCode } from "../types/Op";
-import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
-import { CrdtType } from "../types/SerializedCrdt";
+import { OpCode } from "../protocol/Op";
+import type { IdTuple, SerializedCrdt } from "../protocol/SerializedCrdt";
+import { CrdtType } from "../protocol/SerializedCrdt";
 import { WebsocketCloseCodes } from "../types/WebsocketCloseCodes";
 import {
   listUpdate,
