+++ packages/liveblocks-core/src/__tests__/LiveObject.test.ts
@@ -1,9 +1,9 @@
 import { LiveList } from "..";
 import { RoomScope } from "../AuthToken";
 import { LiveObject } from "../LiveObject";
-import { OpCode } from "../types/Op";
-import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
-import { CrdtType } from "../types/SerializedCrdt";
+import { OpCode } from "../protocol/Op";
+import type { IdTuple, SerializedCrdt } from "../protocol/SerializedCrdt";
+import { CrdtType } from "../protocol/SerializedCrdt";
 import { WebsocketCloseCodes } from "../types/WebsocketCloseCodes";
 import {
   createSerializedList,
