+++ packages/liveblocks-core/src/__tests__/doc.test.ts
@@ -1,7 +1,7 @@
 import type { LiveList } from "../LiveList";
 import type { LiveMap } from "../LiveMap";
 import type { LiveObject } from "../LiveObject";
-import { OpCode } from "../types/Op";
+import { OpCode } from "../protocol/Op";
 import {
   createSerializedList,
   createSerializedMap,
