+++ packages/liveblocks-core/src/protocol/BaseUserMeta.ts
